+++ noodles-vcf/src/lib.rs
@@ -22,13 +22,16 @@
 #[cfg(feature = "async")]
 pub mod r#async;
 
+pub mod fs;
 pub mod header;
-mod indexer;
 pub mod io;
 pub mod record;
 pub mod variant;
 
-pub use self::{header::Header, indexer::index, record::Record};
+pub use self::{header::Header, record::Record};
+
+#[deprecated(since = "0.73.0", note = "Use `vcf::fs::index` instead.")]
+pub use self::fs::index;
 
 #[cfg(feature = "async")]
 pub use self::r#async::io::{Reader as AsyncReader, Writer as AsyncWriter};